requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.9.2/constraints-3.11.txt"

apache-airflow-providers-snowflake==5.5.1
apache-airflow-providers-mysql==5.6.1
pandas
pyyaml
numpy
wget
pybiomart
openpyxl
pytest
boto3
requests
markupsafe
Jinja2
mysqlclient
PyYAML
dsnparse